# docs/swagger.yaml
basePath: /api
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
  food.CreateFoodRequest:
    properties:
      image:
        type: string
      menu_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
    required:
    - name
    - price
    type: object
  food.UpdateFoodRequest:
    properties:
      image:
        type: string
      menu_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
    required:
    - name
    - price
    type: object
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      status:
        type: boolean
    type: object
  user.UpdateUserRequest:
    properties:
      avatar:
        type: string
      email:
        minLength: 3
        type: string
      first_name:
        minLength: 3
        type: string
      last_name:
        minLength: 3
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - phone
    type: object
host: localhost:3000
info:
  contact: {}
  description: restaurant description
  title: Restaurant Management API
  version: "1.0"
paths:
  /auth/login:
    post:
      parameters:
      - description: login body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login
      tags:
      - Auth
  /auth/refresh-token:
    post:
      parameters:
      - description: refrehtoken body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh Token
      tags:
      - Auth
  /auth/signup:
    post:
      parameters:
      - description: signup body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Sign Up
      tags:
      - Auth
  /foods:
    get:
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit NUmber
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all foods
      tags:
      - Foods
    post:
      parameters:
      - description: food data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/food.CreateFoodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create Food
      tags:
      - Foods
  /foods/{id}:
    get:
      parameters:
      - description: food_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get single food by id
      tags:
      - Foods
    patch:
      parameters:
      - description: food data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/food.UpdateFoodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update Food by ID
      tags:
      - Foods
  /users:
    get:
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit number
        in: query
        name: limit
        type: integer
      - description: search user by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: List User
      tags:
      - Users
  /users/{id}:
    get:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get User By ID
      tags:
      - Users
    patch:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: user body update
        in: body
        name: requrest
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update User
      tags:
      - Users
swagger: "2.0"